Noticed that the Dockerfile is using an old version of stretch. Current version is 9.13 not 9.5.
This PR adds dependabot support for updating Dockerfile as well as converting the existing Dockerfile to use a digest for the debian:9.5-slim image.

Once this PR is accepted and merged, dependabot should create a new PR to update to debian:9.13-slim.

However, given that support for Stretch ends in about 5 months according to https://endoflife.date/debian, consideration should be given towards a new Debian version but I will leave that to the maintainers to decide.