:book: Fix recommended command to test the image in development (#1583)
Because no testdata files are actually copied to the image, we fail to
read the event file and fallback to fetching the information from GitHub
API.

Howver, that's not actually working because the `GITHUB_API_URL`
environment is missing.

To fix it, we could pass the `GITHUB_API_URL`. However, I think actually
making the testdata file available to the container mimics better what
happens in real life.

Signed-off-by: David Rodríguez <2887858+deivid-rodriguez@users.noreply.github.com>